requirements.txt
streamlit>=1.30
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.7
seaborn>=0.12
pillow>=10.0
